tratamento de exceções

+++ src/main/java/com/sumauma/dscommerce/services/ProductService.java
@@ -1,14 +1,20 @@
 package com.sumauma.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.sumauma.dscommerce.dto.ProductDTO;
 import com.sumauma.dscommerce.entities.Product;
 import com.sumauma.dscommerce.repositories.ProductRepository;
+import com.sumauma.dscommerce.services.customExceptions.DataBaseException;
+import com.sumauma.dscommerce.services.customExceptions.RessourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -30,7 +36,8 @@ public class ProductService {
 		//return new ProductDTO(result.get());
 		
 		//simplificado
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id).orElseThrow(
+				() -> new RessourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(product);
 							
 	}
@@ -61,16 +68,29 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}
+		catch (EntityNotFoundException e) {
+			throw new RessourceNotFoundException("Recurso não encontrado");
+		}
 		
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new RessourceNotFoundException("recurso não encontrado");
+		}
+		try {
 		repository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DataBaseException("Falha de integridade referencial");
+		}
 
 	}
